requirements.txt
absl-py==1.3.0
appnope==0.1.2
asgiref==3.5.2
astroid==2.6.6
astunparse==1.6.3
backcall==0.2.0
backports.entry-points-selectable==1.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.0.4
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
debugpy==1.5.0
decorator==5.1.0
distlib==0.3.3
Django==4.0.5
entrypoints==0.3
filelock==3.3.1
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
getch==1.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.2
image==1.5.33
importlib-metadata==5.0.0
inflection==0.5.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.0.1
jupyter-client==7.0.6
jupyter-core==4.8.1
keras==2.10.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
libclang==14.0.6
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mccabe==0.6.1
more-itertools==8.8.0
MouseInfo==0.1.3
nest-asyncio==1.5.1
numba==0.56.4
numpy==1.21.1
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.4.0
prompt-toolkit==3.0.20
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.10.0
pylint==2.9.6
PyMsgBox==1.0.9
pyobjc==9.0
pyobjc-core==9.0
pyobjc-framework-Accessibility==9.0
pyobjc-framework-Accounts==9.0
pyobjc-framework-AddressBook==9.0
pyobjc-framework-AdServices==9.0
pyobjc-framework-AdSupport==9.0
pyobjc-framework-AppleScriptKit==9.0
pyobjc-framework-AppleScriptObjC==9.0
pyobjc-framework-ApplicationServices==9.0
pyobjc-framework-AppTrackingTransparency==9.0
pyobjc-framework-AudioVideoBridging==9.0
pyobjc-framework-AuthenticationServices==9.0
pyobjc-framework-AutomaticAssessmentConfiguration==9.0
pyobjc-framework-Automator==9.0
pyobjc-framework-AVFoundation==9.0
pyobjc-framework-AVKit==9.0
pyobjc-framework-BusinessChat==9.0
pyobjc-framework-CalendarStore==9.0
pyobjc-framework-CallKit==9.0
pyobjc-framework-CFNetwork==9.0
pyobjc-framework-ClassKit==9.0
pyobjc-framework-CloudKit==9.0
pyobjc-framework-Cocoa==9.0
pyobjc-framework-Collaboration==9.0
pyobjc-framework-ColorSync==9.0
pyobjc-framework-Contacts==9.0
pyobjc-framework-ContactsUI==9.0
pyobjc-framework-CoreAudio==9.0
pyobjc-framework-CoreAudioKit==9.0
pyobjc-framework-CoreBluetooth==9.0
pyobjc-framework-CoreData==9.0
pyobjc-framework-CoreHaptics==9.0
pyobjc-framework-CoreLocation==9.0
pyobjc-framework-CoreMedia==9.0
pyobjc-framework-CoreMediaIO==9.0
pyobjc-framework-CoreMIDI==9.0
pyobjc-framework-CoreML==9.0
pyobjc-framework-CoreMotion==9.0
pyobjc-framework-CoreServices==9.0
pyobjc-framework-CoreSpotlight==9.0
pyobjc-framework-CoreText==9.0
pyobjc-framework-CoreWLAN==9.0
pyobjc-framework-CryptoTokenKit==9.0
pyobjc-framework-DeviceCheck==9.0
pyobjc-framework-DictionaryServices==9.0
pyobjc-framework-DiscRecording==9.0
pyobjc-framework-DiscRecordingUI==9.0
pyobjc-framework-DiskArbitration==9.0
pyobjc-framework-DVDPlayback==9.0
pyobjc-framework-EventKit==9.0
pyobjc-framework-ExceptionHandling==9.0
pyobjc-framework-ExecutionPolicy==9.0
pyobjc-framework-ExternalAccessory==9.0
pyobjc-framework-FileProvider==9.0
pyobjc-framework-FileProviderUI==9.0
pyobjc-framework-FinderSync==9.0
pyobjc-framework-FSEvents==9.0
pyobjc-framework-GameCenter==9.0
pyobjc-framework-GameController==9.0
pyobjc-framework-GameKit==9.0
pyobjc-framework-GameplayKit==9.0
pyobjc-framework-ImageCaptureCore==9.0
pyobjc-framework-IMServicePlugIn==9.0
pyobjc-framework-InputMethodKit==9.0
pyobjc-framework-InstallerPlugins==9.0
pyobjc-framework-InstantMessage==9.0
pyobjc-framework-Intents==9.0
pyobjc-framework-IOSurface==9.0
pyobjc-framework-iTunesLibrary==9.0
pyobjc-framework-KernelManagement==9.0
pyobjc-framework-LatentSemanticMapping==9.0
pyobjc-framework-LaunchServices==9.0
pyobjc-framework-libdispatch==9.0
pyobjc-framework-LinkPresentation==9.0
pyobjc-framework-LocalAuthentication==9.0
pyobjc-framework-MapKit==9.0
pyobjc-framework-MediaAccessibility==9.0
pyobjc-framework-MediaLibrary==9.0
pyobjc-framework-MediaPlayer==9.0
pyobjc-framework-MediaToolbox==9.0
pyobjc-framework-Metal==9.0
pyobjc-framework-MetalKit==9.0
pyobjc-framework-MetalPerformanceShaders==9.0
pyobjc-framework-MetalPerformanceShadersGraph==9.0
pyobjc-framework-MLCompute==9.0
pyobjc-framework-ModelIO==9.0
pyobjc-framework-MultipeerConnectivity==9.0
pyobjc-framework-NaturalLanguage==9.0
pyobjc-framework-NetFS==9.0
pyobjc-framework-Network==9.0
pyobjc-framework-NetworkExtension==9.0
pyobjc-framework-NotificationCenter==9.0
pyobjc-framework-OpenDirectory==9.0
pyobjc-framework-OSAKit==9.0
pyobjc-framework-OSLog==9.0
pyobjc-framework-PassKit==9.0
pyobjc-framework-PencilKit==9.0
pyobjc-framework-Photos==9.0
pyobjc-framework-PhotosUI==9.0
pyobjc-framework-PreferencePanes==9.0
pyobjc-framework-PushKit==9.0
pyobjc-framework-Quartz==9.0
pyobjc-framework-QuickLookThumbnailing==9.0
pyobjc-framework-ReplayKit==9.0
pyobjc-framework-SafariServices==9.0
pyobjc-framework-SceneKit==9.0
pyobjc-framework-ScreenSaver==9.0
pyobjc-framework-ScreenTime==9.0
pyobjc-framework-ScriptingBridge==9.0
pyobjc-framework-SearchKit==9.0
pyobjc-framework-Security==9.0
pyobjc-framework-SecurityFoundation==9.0
pyobjc-framework-SecurityInterface==9.0
pyobjc-framework-ServiceManagement==9.0
pyobjc-framework-Social==9.0
pyobjc-framework-SoundAnalysis==9.0
pyobjc-framework-Speech==9.0
pyobjc-framework-SpriteKit==9.0
pyobjc-framework-StoreKit==9.0
pyobjc-framework-SyncServices==9.0
pyobjc-framework-SystemConfiguration==9.0
pyobjc-framework-SystemExtensions==9.0
pyobjc-framework-UniformTypeIdentifiers==9.0
pyobjc-framework-UserNotifications==9.0
pyobjc-framework-UserNotificationsUI==9.0
pyobjc-framework-VideoSubscriberAccount==9.0
pyobjc-framework-VideoToolbox==9.0
pyobjc-framework-Virtualization==9.0
pyobjc-framework-Vision==9.0
pyobjc-framework-WebKit==9.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.1
pyzmq==22.3.0
Quandl==3.6.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typing_extensions==4.4.0
urllib3==1.26.6
virtualenv==20.8.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.10.0
